Arduino middle man L1-L2 1st version upload

+++ MultipleLayers/Layer2_ED_Layer1_COOR/Network.h
@@ -16,7 +16,7 @@ class Network{
     int get_data_size(); // get number of extensometer lengths per ED
     int get_payload_size(); // get total length of payload array
     void update_data(uint32_t address, vector<uint16_t> new_data);
-    void construct_payload(uint16_t* payload);
+    void construct_payload(uint8_t* payload);
     private:
     EndDevice* find_device(uint32_t address);
 };
